refactor: split rival viewmodel support responsibilities
