+++ dogArea/Views/ProfileSettingView/RivalTabViewModelSupport/RivalTabViewModel+ModerationAndLocation.swift
@@ -1,0 +1,149 @@
+import SwiftUI
+import CoreLocation
+#if canImport(UIKit)
+import UIKit
+#endif
+
+extension RivalTabViewModel {
+    /// 익명 코드 숨김을 적용하고 즉시 목록에서 제거합니다.
+    func hideAlias(aliasCode: String) {
+        var next = Set(hiddenAliases)
+        next.insert(aliasCode)
+        hiddenAliases = next.sorted()
+        persistModerationPreferences()
+        appendModerationLog(action: "hide", aliasCode: aliasCode, reason: nil)
+        applyLeaderboardModerationFilter()
+        showToast("\(aliasCode) 숨김 처리됐어요")
+    }
+
+    /// 익명 코드 차단을 적용하고 즉시 목록에서 제거합니다.
+    func blockAlias(aliasCode: String) {
+        var blocked = Set(blockedAliases)
+        blocked.insert(aliasCode)
+        blockedAliases = blocked.sorted()
+        var hidden = Set(hiddenAliases)
+        hidden.insert(aliasCode)
+        hiddenAliases = hidden.sorted()
+        persistModerationPreferences()
+        appendModerationLog(action: "block", aliasCode: aliasCode, reason: nil)
+        applyLeaderboardModerationFilter()
+        showToast("\(aliasCode) 차단 처리됐어요")
+    }
+
+    /// 숨김된 익명 코드를 다시 표시 대상으로 복구합니다.
+    func unhideAlias(aliasCode: String) {
+        hiddenAliases.removeAll { $0 == aliasCode }
+        persistModerationPreferences()
+        applyLeaderboardModerationFilter()
+    }
+
+    /// 차단된 익명 코드를 다시 표시 대상으로 복구합니다.
+    func unblockAlias(aliasCode: String) {
+        blockedAliases.removeAll { $0 == aliasCode }
+        hiddenAliases.removeAll { $0 == aliasCode }
+        persistModerationPreferences()
+        applyLeaderboardModerationFilter()
+    }
+
+    /// 신고 사유를 로컬 로그에 남기고 중복 신고를 방지합니다.
+    func reportAlias(aliasCode: String, reason: RivalReportReason) {
+        appendModerationLog(action: "report", aliasCode: aliasCode, reason: reason.rawValue)
+        showToast("\(aliasCode) 신고가 접수됐어요")
+    }
+
+    /// 권한 안내 카드에서 시스템 설정 화면을 엽니다.
+    func openSystemSettings() {
+#if canImport(UIKit)
+        guard let url = URL(string: UIApplication.openSettingsURLString),
+              UIApplication.shared.canOpenURL(url) else {
+            return
+        }
+        UIApplication.shared.open(url)
+#endif
+    }
+
+    /// 짧은 사용자 피드백 메시지를 노출합니다.
+    func showToast(_ message: String) {
+        toastMessage = message
+    }
+
+    /// 노출 중인 토스트를 제거합니다.
+    func clearToast() {
+        toastMessage = nil
+    }
+
+    /// 저장된 숨김/차단 익명 코드 목록을 불러옵니다.
+    func loadModerationPreferences() {
+        let snapshot = moderationStore.loadSnapshot()
+        hiddenAliases = snapshot.hiddenAliases
+        blockedAliases = snapshot.blockedAliases
+    }
+
+    /// 현재 숨김/차단 익명 코드 목록을 로컬 설정에 저장합니다.
+    func persistModerationPreferences() {
+        moderationStore.saveSnapshot(
+            RivalModerationSnapshot(
+                hiddenAliases: hiddenAliases,
+                blockedAliases: blockedAliases
+            )
+        )
+    }
+
+    /// 리더보드 원본 데이터에 숨김/차단 필터를 적용해 사용자 노출 목록을 갱신합니다.
+    func applyLeaderboardModerationFilter() {
+        let blocked = Set(blockedAliases)
+        let hidden = Set(hiddenAliases)
+        leaderboardEntries = latestRawLeaderboardEntries.filter { row in
+            blocked.contains(row.aliasCode) == false && hidden.contains(row.aliasCode) == false
+        }
+        if compareScope == .friend {
+            leaderboardState = .friendPreview
+        } else if leaderboardEntries.isEmpty {
+            leaderboardState = .empty
+        } else {
+            leaderboardState = .ready
+        }
+    }
+
+    /// 신고/차단/숨김 이력을 로컬 JSON 로그에 누적합니다.
+    func appendModerationLog(action: String, aliasCode: String, reason: String?) {
+        moderationStore.appendLog(action: action, aliasCode: aliasCode, reason: reason)
+    }
+
+    /// 위치 권한이 바뀌면 화면 상태를 즉시 재계산합니다.
+    func locationManagerDidChangeAuthorization(_ manager: CLLocationManager) {
+        RivalCoreLocationCallTracer.record(
+            "locationManagerDidChangeAuthorization",
+            detail: "status=\(manager.authorizationStatus.rawValue)"
+        )
+        updatePermissionState()
+        switch manager.authorizationStatus {
+        case .authorizedAlways, .authorizedWhenInUse:
+            RivalCoreLocationCallTracer.record(
+                "startUpdatingLocation",
+                detail: "source=locationManagerDidChangeAuthorization"
+            )
+            manager.startUpdatingLocation()
+        default:
+            RivalCoreLocationCallTracer.record(
+                "stopUpdatingLocation",
+                detail: "source=locationManagerDidChangeAuthorization"
+            )
+            manager.stopUpdatingLocation()
+        }
+        refreshViewState()
+        refreshHotspots(force: true)
+        refreshLeaderboard(force: true)
+    }
+
+    /// 새 좌표를 받으면 공유 상태에서만 핫스팟을 즉시 갱신합니다.
+    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
+        RivalCoreLocationCallTracer.record(
+            "didUpdateLocations",
+            detail: "count=\(locations.count)"
+        )
+        guard locations.isEmpty == false,
+              locationSharingEnabled else { return }
+        refreshHotspots(force: false)
+    }
+}
