+++ scripts/ios_pr_check.sh
@@ -50,6 +50,7 @@ swift scripts/rival_auth_session_guard_unit_check.swift
 swift scripts/rival_auth_session_sync_unit_check.swift
 swift scripts/rival_cllocation_delegate_preconcurrency_unit_check.swift
 swift scripts/rival_mainactor_callback_unit_check.swift
+swift scripts/rival_viewmodel_support_split_unit_check.swift
 swift scripts/season_anti_farming_unit_check.swift
 swift scripts/season_comeback_catchup_unit_check.swift
 swift scripts/season_stage2_pipeline_unit_check.swift
